Create office2011_uninstaller.sh
Using Dockutil to work with the dock and uses the official MS knowledge base article as a guide for what to delete and change. I also assume that you're using a default user profile in /System/Library. I dont want to mess with anyone's profiles so if you're using that to install Office I wont touch them.